refactor AdbHelper
Summary:
Rip out AdbHelper, and its options from the cli package and move it to
appropriate packages, to simplify its usage in other packages.  No
functional changes are intended.

Test plan:
Updated AdbHelperTest to reflect the new structure.
buck test
