+++ src/com/facebook/buck/cli/BUCK
@@ -1,8 +1,5 @@
 EVENTS_SRCS = [
   'CommandEvent.java',
-  'InstallEvent.java',
-  'UninstallEvent.java',
-  'StartActivityEvent.java',
 ]
 java_library(
   name = 'events',
@@ -70,6 +67,7 @@ java_immutables_library(
     ':config',
     ':events',
     '//src/com/facebook/buck/android:exceptions',
+    '//src/com/facebook/buck/android:helpers',
     '//src/com/facebook/buck/android:packageable',
     '//src/com/facebook/buck/android:rules',
     '//src/com/facebook/buck/android:utils',
