+++ test/com/facebook/buck/android/BUCK
@@ -43,6 +43,7 @@ java_library(
     '//test/com/facebook/buck/java:testutil',
     '//test/com/facebook/buck/rules:testutil',
     '//test/com/facebook/buck/testutil:testutil',
+    '//third-party/java/ddmlib:ddmlib',
     '//third-party/java/guava:guava',
     '//third-party/java/junit:junit',
     '//third-party/java/jsr:jsr305',
@@ -62,6 +63,7 @@ java_test(
     ':testutil',
     '//src/com/facebook/buck/android/aapt:aapt',
     '//src/com/facebook/buck/android:exceptions',
+    '//src/com/facebook/buck/android:helpers',
     '//src/com/facebook/buck/android:packageable',
     '//src/com/facebook/buck/android:rules',
     '//src/com/facebook/buck/android:split_dex',
@@ -98,6 +100,7 @@ java_test(
     '//src/com/facebook/buck/util:io',
     '//src/com/facebook/buck/util:util',
     '//src/com/facebook/buck/util/environment:platform',
+    '//third-party/java/args4j:args4j',
     '//test/com/facebook/buck/cli:FakeBuckConfig',
     '//test/com/facebook/buck/java:testutil',
     '//test/com/facebook/buck/event:testutil',
@@ -111,6 +114,7 @@ java_test(
     '//test/com/facebook/buck/testutil:testutil',
     '//test/com/facebook/buck/testutil/integration:integration',
     '//third-party/java/asm:asm',
+    '//third-party/java/ddmlib:ddmlib',
     '//third-party/java/easymock:easymock',
     '//third-party/java/guava:guava',
     '//third-party/java/hamcrest:hamcrest',
